Seperate tiles and built board, render tiles onto webpage
Tiles are are more distinct objects, can now get and set a set of tiles without the context of being organized into a board.
The build function also generates a set of elements onto the webpage, attatching a click listener that grabs the specific tile object.
build will soon become the new render.